refactor(frontend): extract my-groups/observe surface from main.ts (#312)
Continues the conservative split: pull the lead-side observe surface (#201e)
— the "My groups" modal (groups the user leads + members' observable
sessions) and the read-only observe terminal modal — into myGroups.ts
(~280 lines).

Same TDZ-safe pattern as admin.ts: DOM elements re-queried locally; the
main.ts imports (showToast, currentFontSize) used only inside functions; the
4 shared modal consts the Escape guard / applyAdminVisibility reference stay
in main.ts; main.ts imports the two close* functions its registry calls.

main.ts: 4324 → 4070. tsc + Biome clean, vite build green, frontend tests
pass (66).
